python/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.20.0
pyarrow>=10.0.0
plotly>=5.0.0
pyyaml>=6.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
